av1: cleanup

+++ modules/av1/av1.h
@@ -40,6 +40,6 @@ struct obu_hdr {
 
 int    av1_leb128_encode(struct mbuf *mb, size_t value);
 size_t av1_leb128_decode(struct mbuf *mb);
-int    av1_obu_encode(struct mbuf *mb, unsigned type, bool has_size,
+int    av1_obu_encode(struct mbuf *mb, uint8_t type, bool has_size,
 		      size_t len, const uint8_t *payload);
 int    av1_obu_decode(struct obu_hdr *hdr, struct mbuf *mb);
